requirements.txt
flask==2.3.0
qrcode==7.4.2
pillow==10.4.0
Flask-SocketIO==5.3.6
Flask-SQLAlchemy==3.1.1
pytest
PyYAML==6.0.1